requirements.txt
streamlit
pdf2image
python-poppler
boto3
streamlit-cognito-auth
python-dotenv